[FEAT]: added protectedRoutes
